Sankey hover logic in separate composable
